updates and small fixes for inspector and style editor
